+++ spec/rubocop/cli_spec.rb
@@ -1518,7 +1518,7 @@ RSpec.describe RuboCop::CLI, :isolated_environment do
         expect($stdout.string).to eq(<<-RESULT.strip_indent)
             == example/example1.rb ==
             C:  1: 11: Metrics/ParameterLists: Avoid parameter lists longer than 5 parameters. [6/5]
-            C:  1: 39: Naming/UncommunicativeMethodArgName: Method argument must be longer than 3 characters.
+            C:  1: 39: Naming/UncommunicativeMethodParamName: Method parameter must be longer than 3 characters.
             C:  1: 46: Style/CommentedKeyword: Do not place comments on the same line as the def keyword.
             E:  1: 81: Metrics/LineLength: Line is too long. [90/80]